segfault ls->file[i]->name / pathname

+++ include/ft_ls.h
@@ -6,7 +6,7 @@
 /*   By: apouchet <apouchet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/16 11:22:40 by floblanc          #+#    #+#             */
-/*   Updated: 2020/01/24 16:17:37 by apouchet         ###   ########.fr       */
+/*   Updated: 2020/01/24 16:58:47 by apouchet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -95,7 +95,7 @@ typedef struct		s_ls
 {
 	int				flag;
 	char			**to_read;
-	int				size;
+	size_t			size;
 	char			*current_path;
 	t_lf			**file;
 }					t_ls;
